requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
fastapi-mail
python-dotenv
aiosqlite
psycopg2-binary
email-validator
